Add commit_generator.py to the .gitignore file
The commit_generator.py file has been included in the .gitignore file to prevent this file from being tracked by Git. This may indicate that the file is generated locally or used only for development, so it shouldn't be part of version control. A redundant blank line has also been removed.